Add files via upload
Data cube showing individual years between 2002 and 2018 observed with NACO in the L-band. The cube shows a cutout of X3 and is centered in the bright lower star to show the evolution of the X3 system. Above the centered star, a blob is evolving and moves with X3 toward the north. This implies that both objects might be connected. I recommend using QFitsView (https://www.mpe.mpg.de/~ott/QFitsView/) to inspect the file.